pyproject.toml
[project]
name = "llmcouncil"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.77.0",
    "google-generativeai>=0.8.6",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "streamlit>=1.53.1",
]
